join HOST/PORT envs in one *URL env for executor/ml/db services

+++ mindsdb/api/mysql/mysql_proxy/controllers/session_controller.py
@@ -83,23 +83,21 @@ class ServerSessionContorller(SessionController):
     def __init__(self):
         super().__init__()
         self.id = f"session_{uuid4()}"
-        executor_host = os.environ.get("MINDSDB_EXECUTOR_HOSTNAME")
-        executor_port = os.environ.get("MINDSDB_EXECUTOR_PORT")
-        if executor_host and executor_port:
-            self.executor_url = f"http://{executor_host}:{executor_port}"
-        else:
-            self.executor_url = "http://localhost:5500"
-
+        self.executor_url = os.environ.get("MINDSDB_EXECUTOR_URL", None)
+        if self.executor_url is None:
+            raise Exception(f"""{self.__class__.__name__} can be used only in modular mode of MindsDB. 
+                            Use Executor as a service and specify MINDSDB_EXECUTOR_URL env variable""")
         logger.info(
             "%s.__init__: executor url - %s", self.__class__.__name__, self.executor_url
         )
 
     def __del__(self):
         """Terminate the appropriate ServiceSessionController instance as well."""
-        url = self.executor_url + "/" + "session"
-        logger.info(
-            "%s.__del__: delete an appropriate ServiceSessionController with, id - %s on the Executor service side",
-            self.__class__.__name__,
-            self.id,
-        )
-        requests.delete(url, json={"id": self.id})
+        if self.executor_url is not None:
+            url = self.executor_url + "/" + "session"
+            logger.info(
+                "%s.__del__: delete an appropriate ServiceSessionController with, id - %s on the Executor service side",
+                self.__class__.__name__,
+                self.id,
+            )
+            requests.delete(url, json={"id": self.id})
